+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-23 09:52:21
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-08-23 13:38:27
+// Commit: refactor(test): extract reusable atomic widgets from test
 
 class testComponentState {
-  final String commitRef = '3822';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-08-23T09:52:21+05:00');
+  final String commitRef = '3836';
+  final String lastMessage = 'refactor(test): extract reusable atomic widgets from test';
+  final DateTime timestamp = DateTime.parse('2026-08-23T13:38:27+05:00');
 }